GitPush
-Exception Handeling ne cdo shtrese
-Rifaktorim Kodi

+++ RealEstateWebAPI.BLL/Services/PropertiesService.cs
@@ -1,8 +1,9 @@
 ﻿using AutoMapper;
 using RealEstateWebAPI.BLL.DTO;
+using RealEstateWebAPI.Common.ErrorHandeling;
 using RealEstateWebAPI.DAL.Entities;
 using RealEstateWebAPI.DAL.Repositories;
-
+using Serilog;
 
 namespace RealEstateWebAPI.BLL.Services
 {
@@ -11,7 +12,7 @@ namespace RealEstateWebAPI.BLL.Services
         private readonly IPropertyRepository _propertyRepository;
         private readonly IMapper _mapper;
 
-        public PropertiesService(Microsoft.Extensions.Logging.ILogger<PropertiesService> logger, IPropertyRepository propertyRepository, IMapper mapper)
+        public PropertiesService(IPropertyRepository propertyRepository, IMapper mapper)
         {
             _propertyRepository = propertyRepository;
             _mapper = mapper;
@@ -22,16 +23,14 @@ namespace RealEstateWebAPI.BLL.Services
         /// <returns>Nje koleksion Properties pervec atyre qe jane softDelete.</returns>
         public async Task<IEnumerable<PropertyDTO>> GetAllPropertiesAsync()
         {
-            try
-            {
-                var properties = await _propertyRepository.GetAllPropertiesAsync();
 
+            var properties = await _propertyRepository.GetAllPropertiesAsync();
+            if (properties != null)
+            {
+                Log.Information("Got all Properties");
                 return _mapper.Map<IEnumerable<PropertyDTO>>(properties);
             }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            throw new CustomException("Couldnt get Properties");
         }
         /// <summary>
         /// Merr nje property asinkronisht ne baze te id .
@@ -39,37 +38,26 @@ namespace RealEstateWebAPI.BLL.Services
         /// <returns>Nje Property pervec atij qe eshte softDeleted.</returns>
         public async Task<PropertyDTO> GetPropertyByIdAsync(int propertyId)
         {
-            try
+            var property = await _propertyRepository.GetPropertyByIdAsync(propertyId);
+            if (property != null)
             {
-                var property = await _propertyRepository.GetPropertyByIdAsync(propertyId);
+                Log.Information($"Got property with ID: {propertyId}");
                 return _mapper.Map<PropertyDTO>(property);
             }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            throw new CustomException("Property not found");
         }
         /// <summary>
         /// Shton nje Property asinkronisht.
         /// </summary>
-        /// 
-
+        ///
         public async Task<int> AddPropertyAsync(PropertyDTO propertyDTO, int userId)
         {
-            try
-            {
-                var property = _mapper.Map<Property>(propertyDTO);
-                property.UserId = userId; // Assign the user ID to the property
-
-                await _propertyRepository.AddPropertyAsync(property);
-                return property.PropertyId;
-            }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            var property = _mapper.Map<Property>(propertyDTO);
+            property.UserId = userId; // Assign the user ID to the property
+            await _propertyRepository.AddPropertyAsync(property);
+            Log.Information("Property added succesfully");
+            return property.PropertyId;
         }
-
         /* public async Task<int> AddPropertyAsync(PropertyDTO propertyDTO)
          {
              try
@@ -88,22 +76,16 @@ namespace RealEstateWebAPI.BLL.Services
         /// </summary>
         public async Task UpdatePropertyAsync(int propertyId, PropertyDTO propertyDTO, int userId)
         {
-            try
+            var property = await _propertyRepository.GetPropertyByIdAsync(propertyId);
+            if (property != null)
             {
-                var property = await _propertyRepository.GetPropertyByIdAsync(propertyId);
-                if (property != null)
-                {
-                    // Update the property details
-                    _mapper.Map(propertyDTO, property);
-                    property.UserId = userId; // Assign the user ID to the property
-
-                    await _propertyRepository.UpdatePropertyAsync(property);
-                }
+                // Update the property details
+                _mapper.Map(propertyDTO, property);
+                property.UserId = userId; // Assign the user ID to the property
+                await _propertyRepository.UpdatePropertyAsync(property);
+                Log.Information("Property  updated succesfully");
             }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            throw new CustomException("Property not found");
         }
         /* public async Task UpdatePropertyAsync(int propertyId, PropertyDTO propertyDTO)
          {
@@ -125,16 +107,15 @@ namespace RealEstateWebAPI.BLL.Services
         /// Fshin nje Property asinkronisht.
         /// </summary>
         public async Task DeletePropertyAsync(int propertyId)
-
         {
-            try
+            var propertyToDelte = await _propertyRepository.GetPropertyByIdAsync(propertyId);
+            if (propertyToDelte != null)
             {
                 await _propertyRepository.DeletePropertyAsync(propertyId);
+                Log.Information($"Property with id {propertyToDelte.PropertyId} got deleted");
             }
-            catch (Exception ex)
-            {
-                throw;
-            }
+
+            throw new CustomException("Property not found");
         }
         /// <summary>
         /// Merr nje property ne baze te lokacionit asinkronisht.
@@ -142,16 +123,15 @@ namespace RealEstateWebAPI.BLL.Services
         /// <returns>>Nje Property</returns>
         public async Task<IEnumerable<PropertyDTO>> GetAllPropertiesByLocationAsync(string location)
         {
-            try
+
+            var properties = await _propertyRepository.GetPropertyByLocationAsync(location);
+            if (properties != null)
             {
-                var properties = await _propertyRepository.GetPropertyByLocationAsync(location);
+                Log.Information("Got all Properties");
                 return _mapper.Map<IEnumerable<PropertyDTO>>(properties);
             }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            throw new CustomException("Couldnt get properties");
         }
     }
-
 }
+
